[REVERT-ME] meta: packagegroups: multimedia: add libcamera-dev
libcamera recipe on kirkstone is broken because of bad packaging:

libcamera.so.0 and libcamera-base.so.0 are packaged in ${PN}-dev
although they should be packaged in ${PN}. See poky/meta/conf/bitbake.conf.
This trigger the following error when trying to run the `cam` utility
(packaged inside ${PN}):

$ cam -l
cam: error while loading shared libraries: libcamera.so.0: cannot open
shared object file: No such file or directory

$ read-elf -d /usr/bin/cam
Dynamic section at offset 0x2c740 contains 37 entries:
  Tag        Type                         Name/Value
 0x0000000000000001 (NEEDED)             Shared library:
[libcamera.so.0]
 0x0000000000000001 (NEEDED)             Shared library:
[libcamera-base.so.0]
[...]

So package libcamera is broken and need to be installed along with
libcamera-dev to be functionnal.

I sent a patch upstream to fix this [1]. This commit should be reverted
when the upstream patch will be merged.

[1]: https://lists.openembedded.org/g/openembedded-devel/topic/meta_multimedia_kirkstone_patch/91031240?p=,,,20,0,0,0::recentpostdate/sticky,,,20,2,0,91031240,previd%3D1652256959985399985,nextid%3D1652085280557279513&previd=1652256959985399985&nextid=1652085280557279513

Signed-off-by: Julien STEPHAN <jstephan@baylibre.com>