weston-init: Change session user to root
In kirksone, a new user (weston) is created for launching weston server,
which is not convenient for testing purpose. We update bbappend to
change the session user to root, so that we don't need to override
XDG_RUNTIME_DIR before running wayland client.

Also update /home/root/.profile to export WAYLAND_DISPLAY after login.

Bug: AIOT-1070

Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>