distro: rity-demo: add vim support
Currently, the image ships vim-tiny not vim (/usr/bin/vim is a symlink
to /usr/bin/vim.tiny) this is defined inside
poky/.../packagegroup-core-base-utils.bb

Vim fail to source defaults.vim because it is shipped with vim-common
which is not part of vim-tiny.

To solve that, install vim instead of vim-tiny.

Signed-off-by: Alexandre Mergnat <amergnat@baylibre.com>