rng-tools: disabled CPU jitter entropy source in rngd
Add "-x jitter" parameter in rngd init command that the CPU jitter entropy source is disabled.

Note: rngd occupys 4 cpu cores in a short intervel after booting to console.
This is because rngd takes CPU jitter as a RNG entropy source which is based on
running multi-threads on CPUs, and it is tagged as a slow source in rng-tools.

Signed-off-by: Ryan Cho <ryan.cho@mediatek.com>