benchmark-suite: Use prebuilt binaries for tensorflow-lite & glmark2
The benchmark-suite is designed to be built with application SDK, and is
not compatible with Yocto build environment. It's difficult to build
tensorflow-lite (requires Internet access) and glmark2 (without using
meson.bbclass), hense we decide to add support of prebuilt binaries to
the suite, and update the recipe to enable the configuration.

Also remove unneeded patch to add those benchmarks back.

Change-Id: Ie9acbc5b606f0d9e1e50ec58cd2a7c8ea679ac58
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>