base-files: add wwan-5g service for genio-700-evk
Add wwan-5g service for controlling Quectel RM500K 5G card on M.2 slot.
This service depends on usbhub service managed by systemd.

The wwan-5g and usbhub services are used to manage a WWAN 5G card that
is connected to the system via an M.2 slot. The wwan-5g service is
responsible for controlling the WWAN 5G card, while the usbhub service
manages the USB hub that the card is connected to.

Change-Id: I01c5e20dffa335428057ed779b4615849ea607f6
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>