meta: Add new image recipe rity-image-initramfs
The recipe is used for generating initramfs image and to be bundled with
kernel. Currently it does nothing but parses kernel command line,
mounts rootfs and hands over control to the real init.

The recipe is based on core-image-minimal-initramfs.bb, but remove udev
support (which causes size bloated to 13MB). As a remedy, a new initramfs
module is added for finding root device but slimmer.

Change-Id: I5f1126355534fb3e2152000b78378ef37dff1520
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>