meta: u-boot: Load boot script from kernel partition
The u-boot default behavior is to look for boot script in boot
devices. Since we pack boot script in single fitImage file, the beavior
requires a fitImage located under rootfs.

But when INITRAMFS_IMAGE_BUNDLE="1", fitImage won't be installed in
rootfs because of the task order of kernel building (packaging is
performed before generating fitImage), which causes u-boot can not
find boot script and fails.

As a workaround, we overwrite the u-boot env veriable 'mmc_boot' to
load boot script directly from kernel partition instead of rootfs.
The workaround works even if initramfs is disabled.

The long term solution is to separate boot script and fitImage, which
will come in later commits.

Change-Id: I732e8b06d98755a51c633d63528835256e69c734
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>