base-files: genio-1200-evk: add wwan-5g service
Add wwan-5g service for controlling Quectel RM500K 5G card on M.2 slot.
This service depends on usbhub service managed by systemd.

The wwan-5g and usbhub services are used to manage a WWAN 5G card that
is connected to the system via an M.2 slot. The wwan-5g service is
responsible for controlling the WWAN 5G card, while the usbhub service
manages the USB hub that the card is connected to.

Change-Id: Iebeb738fd2b7400d9922da53c4bef2016294df1f
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>