base-files: usbgadget: Generate consistent machine serial
Allow customization of USB serial number, which is also
used by `adb devices`. If not customized, use a hardware id.

Load the U-Boot env variable `serial#`, which can be
programmed by `aiot-flash --serial-no`.

If it is not programmed, read the machine identifier using the optee-ewriter package.
This id would be consistent across multiple image flashing process.

If OPTEE and ewriter are not available, a default serial is being used.

Change-Id: Iaaa2ec8a4461cb2fdad5bdfdbd7a3fb9d089c998
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>