wic-tools: Add python3-pysimg-native as dependency
To generate disk image in sparse image format (for use with fastboot),
we need to add the package to dependencies of wic-tools, so that we can
run the tool (img2simg) without sourcing env file.

The built tool and its dependency are usually located under following paths:
<YOCTO_PROJ>/build/tmp/work/armv8a-poky-linux/wic-tools/1.0-r0/recipe-sysroot-native/usr/bin
and
<YOCTO_PROJ>/build/tmp/work/armv8a-poky-linux/wic-tools/1.0-r0/recipe-sysroot-native/usr/sbin

To use it, add those two paths to the PATH variable.

Change-Id: I3b1bbb6227daf39f6a9216c518b62d1cb45d5456
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>