board-assets: Add custom board name support
To allow an user to specify the custom name of device tree file and
custom search path, we add a mechanism to achieve the task. The user
defines two variable 'MACHINE_DTB' and 'DTB_PATH' in the local.conf,
and the recipe uses these values to modify the content of
u-boot-initial-env, so that u-boot is able to correctly find the
custom device tree files.

We put variable definitions & comments in another file
board-assets-common.inc, which can be shared with other recipes.

Change-Id: I0847a796738ea687225092b57af2b60a06ece906
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>