meta: wic: Partition layout adjustment
To be consistent with EBBR and simplify the partition layout, we remove
the dedicated capsule partition and add an EFI system partition,
though it's not used for booting IOT Yocto. Currently the EFI system
partition contains nothing but the folder structure for capsules used
only for firmware upgrading.

The capsule-part.bb is also updated to reflect the change.

Change-Id: I94e01822744f43a090e89b7ac29b983692e036f1
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>