meta: base-files: usbgadget: fix start-stop-daemon
[Problem]:
By default, usbgadget.service fails to run because start-stop-daemon
requires --exec to be specified with --start and --stop. Also, the udhcpd.conf file needs to be passed as an argument to udhcpd using --. As a result of this, usb and adb are not available.

[Solution]:
Use the correct syntax for start-stop-daemon to get ssh and adb working again.

Change-Id: Ic29907c098d60e63447c7031ba510e4b0e1d711a
Signed-off-by: Suhrid Subramaniam <suhrid.subramaniam@mediatek.com>