meta: Move genio-devtools to the meta folder
The package genio-devtools is referenced by the recipe packagegroup-rity-debug,
which is globally accessible. But genio-devtools is a private recipe of
the layer meta-rity-bsp, and that will bring in unneeded dependency on
the layer when we include packagegroup-rity-debug in other layers. To
fix the issue, we decide to move genio-devtools to the meta folder and
make it globally accessible.

Change-Id: I66655ea6781784f2217a302d593adb412ca3f064
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>