firmware-part: Add multi-devicetree support
Update the shell function 'collect_artifacts' to loop through the
content of KERNEL_DEVICETREE and copy each device tree blob to the
deploy folder.

Please note that the recipe assumes that all board device trees provided
are variants based on the same SOC; if you supply a device tree that is
based on other SOC, it will blindly pack it without complaint. It's
developer's responsibility to make sure the provided dtbs are able to
boot target boards.

Change-Id: I43e581650ec8128b09d16f6c50177ccde0a980ce
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>